+++ yesmate/js/sounds.js
@@ -1,0 +1,9 @@
+import { seesound } from './seesound/index';
+
+export class Index {
+    constructor(){  
+      seesound.createChart()
+    }
+};
+
+document.body.onload = new Index();